+++ app/src/test/java/com/freevibe/ui/screens/settings/SettingsViewModelTest.kt
@@ -182,6 +182,7 @@ class SettingsViewModelTest {
             wallpaperApplier = wallpaperApplier,
             videoWallpaperStorage = videoWallpaperStorage,
             sourceMetrics = com.freevibe.service.SourceMetrics(),
+            ioDispatcher = dispatcher,
         )
     }